[TASK] Work around bug in EventDispatcher
Composer's EventDispatcher keeps track of circular invocations by
adding the currently handled event to an event stack. If a new event is
dispatched and another event of the same type is already present in the
stack, an error is triggered.

If an exception is thrown in an event handler, the internal state of
EventDispatcher is not correctly dealt with causing subsequent events
with the same type as the failed event to now also fail. This commit
works around this behavior by explicitly being verbose about exceptions
that are thrown in the event handler.