[BUGFIX] Support create-project by extracting plugin interface
Composer exposes a peculiar behavior in a create-project command: an
internal restart is triggered to install dependencies during which all
plugins are reloaded. To ensure that there are no stale references in
the plugin instances, the plugin's main class' source code is
retrieved, a random string is appended to the class name and the
resulting code is eval()'ed.

This breaks type hinting with the main class' name. This commit
extracts two methods from VelocitaPlugin into VelocitaPluginInterface,
the latter of which is then used for type hinting.